chore: move helper methods into the service
