+++ api/src/unraid-api/graph/resolvers/customization/customization.resolver.ts
@@ -1,8 +1,5 @@
 import { Query, ResolveField, Resolver } from '@nestjs/graphql';
 
-import { fileExists } from '@app/core/utils/files/file-exists.js'; // Import utility
-
-import { store } from '@app/store/index.js'; // Import store
 import { Public } from '@app/unraid-api/auth/public.decorator.js'; // Import Public decorator
 
 import {
@@ -21,24 +18,6 @@ import { CustomizationService } from '@app/unraid-api/graph/resolvers/customizat
 @Resolver(() => Customization)
 export class CustomizationResolver {
     constructor(private readonly customizationService: CustomizationService) {}
-
-    private async _getPublicPartnerInfoInternal(): Promise<PublicPartnerInfo | null> {
-        const activationData = await this.customizationService.getActivationData();
-
-        return {
-            hasPartnerLogo: (await this.customizationService.getPartnerLogoWebguiPath()) !== null,
-            partnerName: activationData?.partnerName,
-            partnerUrl: activationData?.partnerUrl,
-            partnerLogoUrl: await this.customizationService.getPartnerLogoWebguiPath(),
-        };
-    }
-
-    private async isPasswordSet(): Promise<boolean> {
-        const paths = store.getState().paths;
-        const hasPasswd = await fileExists(paths.passwd);
-        return hasPasswd;
-    }
-
     // Authenticated query
     @Query(() => Customization, { nullable: true })
     @UsePermissions({
@@ -55,15 +34,15 @@ export class CustomizationResolver {
     @Query(() => PublicPartnerInfo, { nullable: true })
     @Public()
     async publicPartnerInfo(): Promise<PublicPartnerInfo | null> {
-        if (!(await this.isPasswordSet())) {
-            return this._getPublicPartnerInfoInternal();
+        if (!(await this.customizationService.isPasswordSet())) {
+            return this.customizationService.getPublicPartnerInfo();
         }
         return null;
     }
 
     @ResolveField(() => PublicPartnerInfo, { nullable: true, name: 'partnerInfo' })
     async resolvePartnerInfo(): Promise<PublicPartnerInfo | null> {
-        return this._getPublicPartnerInfoInternal();
+        return this.customizationService.getPublicPartnerInfo();
     }
 
     @ResolveField(() => ActivationCode, { nullable: true, name: 'activationCode' })
